Move all Prettier logic to the worker
This way the only data that needs to be cloned using the Structured
Clone Algorithm is the string contents. This means the Prettier
configuration may now use plugins.

Because `tsx` was unable to resolve `worker.ts` with a `.js` extension,
the code had to be changed to JavaScript. Because of a bug in
TypeScript, type definitions had to be written manually.

Closes #1